[tests] put back snapshot tests
