+++ packages/mdx-live/package.json
@@ -32,7 +32,7 @@
     },
     "scripts": {
         "build": "rm -rf lib/ && microbundle -f modern,esm",
-        "test": "node --loader=ts-node/esm/transpile-only --test 'src/*.test.*'"
+        "test": "node --loader=ts-node/esm/transpile-only --test 'src/*.test.{ts,tsx,mjs,cjs}'"
     },
     "dependencies": {
         "@types/mdx": "^2.0.3",
